project-3/flask/requirements.txt
flask==2.3.3
flask-cors>=3.0.10
langchain>=0.0.280
openai>=0.11.5
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pymupdf>=1.22.5
